<refactor>[kvm]: reorganize vm host file management packages
Move common vm host file classes from efi to vmfiles package
Move message classes to message subdirectories
Update all import statements to reference new package paths
Update Spring configuration files with new class locations
Update test configuration files to match new package paths

Related: ZSV-11310
