+++ plugin/kvm/src/main/java/org/zstack/kvm/efi/KvmSecureBootManager.java
@@ -47,6 +47,14 @@ import org.zstack.kvm.KVMAgentCommands;
 import org.zstack.kvm.KVMConstant;
 import org.zstack.kvm.KvmCommandSender;
 import org.zstack.kvm.KvmResponseWrapper;
+import org.zstack.kvm.vmfiles.message.BackupVmHostFileMsg;
+import org.zstack.kvm.vmfiles.message.BackupVmHostFileOnHypervisorMsg;
+import org.zstack.kvm.vmfiles.message.BackupVmHostFileOnHypervisorReply;
+import org.zstack.kvm.vmfiles.message.BackupVmHostFileReply;
+import org.zstack.kvm.vmfiles.message.CloneVmHostFileMsg;
+import org.zstack.kvm.vmfiles.message.CloneVmHostFileReply;
+import org.zstack.kvm.vmfiles.message.SyncVmHostFilesFromHostMsg;
+import org.zstack.kvm.vmfiles.message.SyncVmHostFilesFromHostReply;
 import org.zstack.resourceconfig.ResourceConfig;
 import org.zstack.resourceconfig.ResourceConfigFacade;
 import org.zstack.utils.DebugUtils;
