+++ plugin/kvm/src/main/java/org/zstack/kvm/efi/NvRamVmHostBackupFileBase.java
@@ -2,6 +2,7 @@ package org.zstack.kvm.efi;
 
 import org.zstack.header.vm.additions.VmHostBackupFileVO;
 import org.zstack.header.vm.additions.VmHostFileType;
+import org.zstack.kvm.vmfiles.AbstractVmHostBackupFileBase;
 
 public class NvRamVmHostBackupFileBase extends AbstractVmHostBackupFileBase {
     public NvRamVmHostBackupFileBase(VmHostBackupFileVO self) {
