+++ plugin/kvm/src/main/java/org/zstack/kvm/efi/NvRamVmHostFileBase.java
@@ -2,6 +2,7 @@ package org.zstack.kvm.efi;
 
 import org.zstack.header.vm.additions.VmHostFileType;
 import org.zstack.header.vm.additions.VmHostFileVO;
+import org.zstack.kvm.vmfiles.AbstractVmHostFileBase;
 
 public class NvRamVmHostFileBase extends AbstractVmHostFileBase {
     public NvRamVmHostFileBase(VmHostFileVO self) {
