+++ plugin/kvm/src/main/java/org/zstack/kvm/tpm/message/CloneVmTpmReply.java
@@ -1,4 +1,4 @@
-package org.zstack.kvm.tpm;
+package org.zstack.kvm.tpm.message;
 
 import org.zstack.header.message.MessageReply;
 import org.zstack.header.tpm.entity.TpmInventory;
